requirements.txt
pydantic>=2.0
PyYAML>=6.0
numpy>=1.24
rich>=13.0
# 可选: redis>=5.0  (生产环境 RedisBackend)
